refactor(game-context): decompose 846-line monolith into 6 domain hooks
p776h4 — game-context.tsx reduced from 846 to 259 lines (orchestration layer
only). Business logic extracted into web/lib/contexts/:

  useUIActions     — go/setPopup/setMenuOpen/toasts/dismissToast/clearTerritory
  useAuthSync      — PocketBase backend load/save, guest auth, auth gate, upgrade prompt
  useCatalogSync   — fetchCatalog, daily contractor pool, midnight rollover, refinery unlock notification
  useGameLoop      — onPickMission/onPickTarget/onPurchaseRocket/onLaunch/onMiningDone/onDebriefDone
  useTutorialActions — setTutorial/skipTutorial/setDoneSteps/completeStep/coachManualNext
  useEconomyActions  — sellMinerals/refinery/scanner/skills/loan/abandonMission

GameProvider composes the six hooks and assembles the same flat GameState &
GameActions value; useGame() and all consumers are unchanged. TypeScript strict
clean, 166 unit tests passing.

Also adds Cypress regression assertion for h20xtc (map height > 400px on
mobile 390×844 after the flex-fill layout change).
